refactor: prepare environments package
